Map over fields in AddItem rather than hard-code fields
